+++ types-create/types/augmentRegistry.ts
@@ -1,12 +1,12 @@
 // Copyright 2017-2022 @polkadot/types-create authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import 'https://deno.land/x/polkadot@0.0.4-4/types-codec/types/registry.ts';
+import 'https://deno.land/x/polkadot/types-codec/types/registry.ts';
 
-import type { Codec, CodecClass, ICompact, INumber } from 'https://deno.land/x/polkadot@0.0.4-4/types-codec/types/index.ts';
-import type { ILookup, TypeDef } from 'https://deno.land/x/polkadot@0.0.4-4/types-create/types/index.ts';
+import type { Codec, CodecClass, ICompact, INumber } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
+import type { ILookup, TypeDef } from 'https://deno.land/x/polkadot/types-create/types/index.ts';
 
-declare module 'https://deno.land/x/polkadot@0.0.4-4/types-codec/types/registry.ts' {
+declare module 'https://deno.land/x/polkadot/types-codec/types/registry.ts' {
   interface Registry {
     readonly lookup: ILookup;