+++ types/interfaces/utility/types.ts
@@ -1,8 +1,8 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Struct, Vec, u32 } from 'https://deno.land/x/polkadot@0.0.4-4/types-codec/mod.ts';
-import type { AccountId, Balance, BlockNumber } from 'https://deno.land/x/polkadot@0.0.4-4/types/interfaces/runtime/index.ts';
+import type { Struct, Vec, u32 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { AccountId, Balance, BlockNumber } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
 
 /** @name Multisig */
 export interface Multisig extends Struct {